documentation (tests still needed)
