+++ Contuity/Database/DatabaseManager.swift
@@ -1,0 +1,36 @@
+//
+//  DatabaseManager.swift
+//  Contuity
+//
+//  Created by Anand Kumar on 10/16/18.
+//  Copyright © 2018 Generate. All rights reserved.
+//
+
+import SQLite
+
+/// Singleton instance of our DatabaseManager, initialized at runtime.
+class DatabaseManager {
+
+    /// Get shared static instance
+    static let shared = DatabaseManager()
+
+    /// Enforce only one instance
+    private init() {}
+
+    private (set) var conn: Connection?
+
+    /// Entry point to attach a connection. Enforces that only one connection
+    /// can be attached during the lifecycle of the app.
+    ///
+    /// - Parameter conn: a Database connection
+    func attachConnection(_ conn: Connection) {
+        if self.conn == nil {
+            self.conn = conn
+        }
+    }
+
+    /// Table initializer
+    func initializeTables() {
+        try? Jot.createTable()
+    }
+}
